refactor: useDragDrop step 12 — mobile touch end / swipe actions
